refactor model configuration
